test: make lifecycle docker runner script-safe
